+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * docs: improve marketing and llms.txt
 * chore(deps): update dependency reka-ui to v2.7.0
 * feat(InputMenu/SelectMenu): add `clear` prop
+* feat(CommandPalette): add `input` prop
 
 ## [2.2.1](https://github.com/bitrix24/b24ui/compare/v2.1.17...v2.2.1) (2025-12-18)